# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Документация для API бронирования мест",
        "contact": {
            "email": "support@yourdomain.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/shows": {
            "get": {
                "tags": [
                    "Shows"
                ],
                "summary": "Получить список всех мероприятий",
                "operationId": "71aa465375ef08cb18fe0f645b9028c3",
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Concert"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "A great music concert"
                                            },
                                            "date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2024-08-20"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/shows/{id}/events": {
            "get": {
                "tags": [
                    "Shows"
                ],
                "summary": "Получить список событий для указанного мероприятия",
                "operationId": "9fd19ee035588295adbde5ae71c0988b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID мероприятия",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "show_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "date": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-08-20T19:00:00Z"
                                            },
                                            "location": {
                                                "type": "string",
                                                "example": "Main Hall"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/events/{id}/seats": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Получить доступные места для события",
                "operationId": "172c8ef36820bb119c2977143779e819",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID события",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "seat_id": {
                                                "type": "integer",
                                                "example": 101
                                            },
                                            "row": {
                                                "type": "string",
                                                "example": "A"
                                            },
                                            "number": {
                                                "type": "integer",
                                                "example": 12
                                            },
                                            "available": {
                                                "type": "boolean",
                                                "example": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/events/{id}/book": {
            "post": {
                "tags": [
                    "Booking"
                ],
                "summary": "Бронирование мест",
                "operationId": "9b6df7b6adc74a542e27fb7113d8690f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID события",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "places": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешное бронирование",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Seats booked successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Shows",
            "description": "Shows"
        },
        {
            "name": "Events",
            "description": "Events"
        },
        {
            "name": "Booking",
            "description": "Booking"
        }
    ]
}